+++ packages/engine-test-utils/src/__tests__/dendron-cli/commands/doctor.spec.ts
@@ -17,27 +17,7 @@ import _ from "lodash";
 import path from "path";
 import { GitTestUtils, TestConfigUtils } from "../../..";
 import { createEngineFromServer, runEngineTestV5 } from "../../../engine";
-import { checkFile } from "../../../utils";
-
-class TestDoctorUtils {
-  static createNotesWithNoFrontmatter = ({ wsRoot, vaults }: WorkspaceOpts) => {
-    const payload = ["hello"];
-    const vpath = vault2Path({ vault: vaults[0], wsRoot });
-    return fs.writeFile(path.join(vpath, "test.md"), payload.join("\n"));
-  };
-
-  static createNotesWithBadIds = ({ wsRoot, vaults }: WorkspaceOpts) => {
-    const notes = NoteTestUtilsV4.createNote({
-      wsRoot,
-      fname: "test",
-      vault: vaults[0],
-      props: {
-        id: "-bad-id",
-      },
-    });
-    return { notes };
-  };
-}
+import { checkFile, TestDoctorUtils } from "../../../utils";
 
 const setupBasic = async (opts: WorkspaceOpts) => {
   const { wsRoot, vaults } = opts;
